requirements.txt
# Guava Ripeness Classifier - Dependencies
# PyTorch and Deep Learning
torch>=2.3.0
torchvision>=0.18.0

# Data Processing
numpy>=1.26.0
pandas>=2.2.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Machine Learning Utilities
scikit-learn>=1.5.0
tqdm>=4.66.0

# Computer Vision
opencv-python>=4.8.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.0.0

# API/Backend (for future integration)
flask>=3.0.0
flask-cors>=4.0.0

# Graphviz for flowcharts (optional)
graphviz>=0.20.0
